Add files via upload
Topic: Branching Statements and Logical Operators Homework
Ex1:
The user enters a symbol from the keyboard. Determine what a symbol is: letter, number, punctuation mark or something else.
Exercise 2.
Write a program to calculate the cost of calls for different mobile operators. The user enters the time of call and chooses the operator to make a call. Display the cost on the screen.
Exercise 3.
Vasya works as a programmer and gets $ 50 for every 100 strings of code. For every third delay Vasya is fined $ 20. Implement a menu:
the user enters a desired income of Vasey and a number of delays; count the number of strings of code that he should write;
the user enters a number of strings of code written by Vasya and a desired amount of salary. Calculate how many times Vasya could delay;
the user enters a number of strings of code and a number of delays; determine how much money Vasya will earn and whether it is paid.
   1